+++ CHANGELOG.md
@@ -4,8 +4,16 @@ All notable changes to the "vscode-commit-reminder" extension will be documented
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1] Aug 20, 2020
+## [0.3.0] Oct 26, 2020
+### Added
+    - New config option intervalUnit, defaults to minutes
 ### Changed
+    - Renamed output channel from VSCode Commit Reminder > Commit Reminder
+### Misc
+    - Updated dependencies
+
+## [0.2.1] Aug 20, 2020
+### Misc
     - Updated dependencies
 
 ## [0.2.0] Jul 4, 2020